+++ NEWS.md
@@ -3,6 +3,8 @@
   a factor nRecInDays.
 - Plotting daily sums: Uncertainty computation conservatively assumes fully
   correlated errors rather than fully independent errors
+- Provide argument isSplitDatetime to fWriteDataframeToFile to write files
+  consistent with the REddyProc web tool.
 
 # REddyProc 1.3.2
 - Avoid strong dependency on mlegp (https://github.com/bgctw/REddyProc/issues/58).